Some fixse for automatic text generation Lab/VitalSigns/BloodGroup
git-svn-id: https://svn.code.sf.net/p/ehealthconnector/code/trunk@1360 e99adb7c-5e35-42fd-8d1b-cdb378af05e8

+++ ehealth_connector-communication/ehealth_connector-communication-ch/src/main/java/org/ehealth_connector/fhir/ch/AbstractFhirCdaCh.java
@@ -1266,8 +1266,11 @@ public abstract class AbstractFhirCdaCh {
 						fhirObservation.getReferenceRangeFirstRep().getHigh().getValue(),
 						fhirObservation.getReferenceRangeFirstRep().getHigh().getUnit());
 			} else {
-				v = new Value(fhirObservation.getReferenceRangeFirstRep().getLow().getValue(),
-						fhirObservation.getReferenceRangeFirstRep().getHigh().getValue());
+				v = new Value(
+						fhirObservation.getReferenceRangeFirstRep().getLow().getValue()
+								.toBigInteger(),
+						fhirObservation.getReferenceRangeFirstRep().getHigh().getValue()
+								.toBigInteger());
 			}
 
 			rr.setValue(v);
