+++ ehealth_connector-communication/ehealth_connector-communication-ch/src/main/java/org/ehealth_connector/fhir/ch/FhirCdaChLrph.java
@@ -836,8 +836,11 @@ public class FhirCdaChLrph extends AbstractFhirCdaCh {
 						fhirObservation.getReferenceRangeFirstRep().getHigh().getValue(),
 						fhirObservation.getReferenceRangeFirstRep().getHigh().getUnit());
 			} else {
-				v = new Value(fhirObservation.getReferenceRangeFirstRep().getLow().getValue(),
-						fhirObservation.getReferenceRangeFirstRep().getHigh().getValue());
+				v = new Value(
+						fhirObservation.getReferenceRangeFirstRep().getLow().getValue()
+								.toBigInteger(),
+						fhirObservation.getReferenceRangeFirstRep().getHigh().getValue()
+								.toBigInteger());
 			}
 
 			rr.setValue(v);
